requirements.txt
fastapi==0.89.1
pydantic==1.10.4
pyTelegramBotAPI==4.9.0
requests==2.28.2
pytz~=2022.7.1
python-dotenv~=0.21.0